# backend/src/main/resources/public/openapi.yaml
openapi: 3.0.3
info:
  title: Analytics Trainer API
  version: 1.0.0
servers:
  - url: http://localhost:8080
tags:
  - name: System
    description: Health, metrics and service metadata
  - name: Auth
    description: OTP registration, login and token issuing
  - name: Profile
    description: Current user profile
  - name: Trainers
    description: Trainer catalog and task lists
  - name: Tasks
    description: Task details and answer submission
  - name: Progress
    description: User progress and attempts
paths:
  /health:
    get:
      tags: [System]
      responses:
        "200":
          description: Service health
  /metrics:
    get:
      tags: [System]
      responses:
        "200":
          description: Prometheus metrics
  /api/auth/register/request:
    post:
      tags: [Auth]
      summary: Request email OTP registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, fullName]
              properties:
                email: { type: string }
                password: { type: string }
                fullName: { type: string }
      responses:
        "202": { description: OTP sent }
  /api/auth/register/confirm:
    post:
      tags: [Auth]
      summary: Confirm OTP and create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, otp]
              properties:
                email: { type: string }
                otp: { type: string }
      responses:
        "200": { description: JWT and user }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200": { description: JWT and user }
  /api/me:
    get:
      tags: [Profile]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current user }
  /api/trainers:
    get:
      tags: [Trainers]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Trainer list }
  /api/trainers/{id}/tasks:
    get:
      tags: [Trainers]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Task list }
  /api/tasks/{id}:
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Task details }
  /api/tasks/{id}/submit:
    post:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answer:
                  oneOf:
                    - type: array
                      items: { type: string }
                    - type: object
                    - type: string
      responses:
        "200": { description: Score result }
  /api/progress:
    get:
      tags: [Progress]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: User progress }
  /api/attempts:
    get:
      tags: [Progress]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Attempts }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token